refactor: move plugin/sampling/leaderelection to internal/leaderelection
