+++ internal/leaderelection/mocks/ElectionParticipant.go
